requirements.txt
streamlit>=1.35.0
pandas>=2.2.0
numpy>=1.26.0
sqlalchemy>=2.0.25
openpyxl>=3.1.2
xlsxwriter>=3.1.9
plotly>=5.17.0
pillow>=10.0.1
python-dateutil>=2.8.2
bcrypt>=4.0.1
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
qrcode[pil]>=7.4.2
phonenumbers>=8.13.22
email-validator>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
twilio>=8.8.0
reportlab>=4.0.4
matplotlib>=3.7.2
seaborn>=0.12.2
streamlit-option-menu>=0.3.6